+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/scheduling/ScheduledBeanConfiguration.groovy
@@ -1,0 +1,14 @@
+package com.ofg.infrastructure.scheduling
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class ScheduledBeanConfiguration {
+
+    @Bean
+    TestBeanWithScheduledMethod testBeanWithScheduledMethod() {
+        return new TestBeanWithScheduledMethod()
+    }
+
+}