+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/scheduling/TestBeanWithScheduledMethod.groovy
@@ -1,0 +1,15 @@
+package com.ofg.infrastructure.scheduling
+
+import com.ofg.infrastructure.correlationid.CorrelationIdHolder
+import org.springframework.scheduling.annotation.Scheduled
+
+class TestBeanWithScheduledMethod {
+
+    String correlationId
+
+    @Scheduled(fixedDelay=100L)
+    void scheduledMethod() {
+        correlationId = CorrelationIdHolder.get()
+    }
+
+}