+++ src/main/groovy/com/ofg/stub/StubRunning.groovy
@@ -1,0 +1,5 @@
+package com.ofg.stub
+
+interface StubRunning {
+    void runStubs()
+}