+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 New features:
 * [Issue 18](https://github.com/4finance/micro-infra-spring/issues/18) Add correlationId to headers of a JMS message
+* [Issue 61](https://github.com/4finance/micro-infra-spring/issues/61) When action is triggered by a scheduler CorrelationID is NOT SET
 * CSV metrics publisher
 
 Notable changes: